refactor: use Discord SlashCommandBuilder
In order to follow the recommended way of installing commands, we need to use the builder package,
so convert all commands over to use that.

Adds an install script that uses the rest API to install.

Cleanup config.

BREAKING CHANGE: commands are not installed automatically as part of the bot. You will need to run
the install script