refactor(logger): upgrade to pino 7
Pino 7 removes the pretty print option. The recommended way to pretty print is using pino-pretty as
the script. This requires that the output be stdout, not stderr. Remove the Pino types package as
this is now included. The Level is not typed, instead being a string.

BREAKING CHANGE: Removes the APP_DEBUG option